Added project to main branch
Added Game Maker project to main branch for Piratesoftware Game Jam; The theme is "Shadows and Alchemy". Added player sprite and wall sprite + objects for them (placeholder colored blocks with outlines); Added collision, moving, and jumping (subject to change) for player sprite and collision against wall; Modified default room to have walls and player sprite to function as a testing room, so it can be used to test and develop the movement, mechanics and so forth.